requirements.txt
Django==1.3.1
South==0.7.3
django-dajaxice==0.2
feedparser==5.0.1
gunicorn==0.13.4
httplib2==0.7.4
oauth2==1.5.211
psycopg2==2.4.2
python-openid==2.2.5
pytz==2011n
wsgiref==0.1.2
